requirements.txt
adlfs==2023.12.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.31.0
azure-datalake-store==0.0.53
azure-identity==1.19.0
azure-storage-blob==12.23.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dacite==1.8.1
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.5.0
fsspec==2024.10.0
GQLAlchemy==1.6.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
MarkupSafe==3.0.2
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
neo4j==4.4.12
networkx==2.8.8
numpy==1.26.4
openai==1.52.2
portalocker==2.10.1
propcache==0.2.0
psutil==5.9.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymgclient==1.3.1
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.16.0
zipp==3.20.2
